fix various HostApplication issues around setComponentHandler() and terminate().
The previous change did not really detect IEditController class identity and
always treated as if it were from the same instance, which was incorrect.
Fixing that issue uncovered the problem that we didn't really fix terminate().

It was mostly due to incorrect IComponentHandler pointer from HostApplication,
and fixing its vtable uncovered the issue that we release HostApplication
every time `createInstance()` is completed (as it was a local variable).

Fixing the scope issue fixed most of the issues regarding terminate().

There are some remaining issues that causes some JUCE leaked object (meaning
that we leak some component) and some SIGABRT issues w/ Kontakt and Battery.
There is also some issue that iZotope plugins somehow requires libraries
retained at `AudioPluginLibraryPool`. But they can instantiate and terminate.