Reorganize the relationship between track and plugins. (Codex)
The old model was that an AudioPluginTrack is mapped to a virtual UMP device
and contains one or more plugins. But that means we only have one MIDI device
for the entire track, which means the effet chains have no MIDI support
(in particular, no properties such as AllCtrlList and ProgramList).

The new model flips the relationship between the virtual MIDI device and
tracks. Now a track can contain multiple virtual UMP devices, and each of
the device maps to an audio plugin instance.