redefine PluginParameter and PluginParameterSupport API to treat plain values.
It was always confusing between normalized and plain values in our APIs,
which sometimes brought in confusion. Now, these APIs always deal with
plain values, which match AU, LV2, and CLAP (not VST3).

Now setParameterValue() and getParameterValue() deal with plain value too.

It works most cases, but we have to deal with "relative NRPN" case, where
the value calculation needs to be done in plain value but then normalized
between UINT32 integer range.