Save only DSP states. This fixes some VST3 state loader issues. (Claude)
The loader expected that the state always began with `RST3` (that either
Claude or Codex hacked the idea) while it is saved only if the DSP and
the edit controller both had the state for each. sforzando, for example,
didn't have the GUI state (good!) and failed to restore the state.